requirements.txt

asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
django-multiselectfield==0.1.12
gunicorn==20.1.0
idna==3.3
instaloader==4.8.2
mysql-connector-python==8.0.28
numpy==1.22.0
pandas==1.3.5
Pillow==9.1.0
protobuf==3.19.3
psycopg2==2.9.3
python-dateutil==2.8.2
pytube==12.0.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
whitenoise==5.3.0
